approve status

+++ src/Pages/ManageOrders/ManageOrders.js
@@ -2,58 +2,71 @@ import React, { useEffect, useState } from 'react';
 
 const ManageOrders = () => {
     const [manageOrders, setManageOrders] = useState([])
-    const [isDeleted, setIsDeleted] = useState(null)
-    const [approve, setApprove] = useState('Pending')
+    // const [isDeleted, setIsDeleted] = useState(null)
+    // const [approve, setApprove] = useState('Pending')
     // const [isUpdate, setIsUpdate] = useState(null)
-    console.log(isDeleted)
+    // console.log(isDeleted)
 
 
     useEffect(() => {
         fetch("https://mysterious-island-82371.herokuapp.com/manageOrders")
         .then(res => res.json())
         .then(result => setManageOrders(result))
-    }, [isDeleted])
+    }, [manageOrders])
 
 
+    // delete button 
     const handleDeleteManageOrders = (id) => {
-        fetch(`http://localhost:5000/deleteManageOrders/${id}`,{
+        const proceed = window.confirm('Are you sure you want to delete')
+        if(proceed){
+            fetch(`http://localhost:5000/deleteManageOrders/${id}`,{
             method: "DELETE",
             headers: {"content-type": "application/json"}
         })
         .then(res => res.json())
         .then(result => {
             if(result.deletedCount){
-                setIsDeleted(true);
-                alert('Are sure you want to Delete?')
-            }
-            else{
-                setIsDeleted(false);
+                
+                alert(' Delete Success')
             }
         })
+
+        }
         
         console.log(id)
     }
-    // Update booked Status
-    // const handleUpdateManageOrders = (id) => {
-    //     fetch(`http://localhost:5000/deleteManageOrders/${id}`,{
-    //         method: "PUT",
-    //         headers: {"content-type": "application/json"}
-    //     })
-    //     .then(res => res.json())
-    //     .then(result => {
-    //         if(result.modifiedCount > 0){
-    //             alert('Approved Success')
-    //         }
-    //     })
-        
-    //     console.log(id)
-    // }
+
+
+
+    // Approved button
+    const handleUpdateManageOrders = (id) => {
+        const updateStatus = manageOrders.find(
+            (update) => update?._id === id
+        );
+        updateStatus.status = "approved";
+
+        const uri = `http://localhost:5000/manageOrderUpdate/${id}`;
+        fetch(uri, {
+            method: "PUT",
+            headers: {
+                "content-type": "application/json",
+            },
+            body: JSON.stringify(updateStatus),
+        })
+            .then((res) => res.json())
+            .then((data) => {
+                if (data.modifiedCount > 0) {
+                    alert("Approved successfully");
+                }
+            });
+    };
+
+    
 
     return (
         <div>
-            <h2>Manage Orders Here {manageOrders.length}</h2>
             <div className="text-center mt-5">
-                <h1 className="my-5">Rooms & suits</h1>
+                <h1 className="my-5">All Booked Room Is Here</h1>
             </div>
             <div className="m-auto mb-5">
                 <div className="row container  m-auto">
@@ -73,13 +86,13 @@ const ManageOrders = () => {
 
                                     <h6 className="">Per Night: ${manageOrder.singleServiceDat.singleService.price}</h6>
 
-                                    <h6 onClick={() => setApprove("Approved")} className="">Approve</h6>
+                                    {/* <h6 onClick={() => setApprove("Approved")} className="">Approve</h6> */}
 
-                                    {/* approve btn
-                                    <button onClick={() => setApprove("Approved")} className="me-2 pe-2">Pending</button> */}
+                                    {/* approve btn */}
+                                    <button onClick={() => handleUpdateManageOrders(manageOrder._id)} className="px-4 mx-1 py-1 my-2">{manageOrder.status}</button>
 
                                     {/* delete btn */}
-                                    <button onClick={() => handleDeleteManageOrders(manageOrder._id)} className="pe-2">Delete</button>
+                                    <button onClick={() => handleDeleteManageOrders(manageOrder._id)} className="px-5 py-1 my-2">Delete</button>
                                 </div>
                             </div>
                         </div>
